refactor(api): convert body validation to computed and drop barrel reexports
- Migrate `bodyResultOf` from per-route `command` factory to a single shared
  `computed` that derives the body schema from `route$`, mirroring `queryOf` /
  `pathParamsOf`. Body parsing now memoizes per request.
- Drop the unused barrel re-export in `signals/external/env.ts` and remove the
  redundant `signals/external/log.ts` re-export file; consumers now import
  `lib/env` and `lib/log` directly.
- Update device-token route handlers to read body via `get(...)` and add
  required `signal.throwIfAborted()` after each await per the
  `signal-check-await` rule.
