+++ turbo/apps/api/src/signals/context/request.ts
@@ -1,5 +1,5 @@
 import type { AppRoute } from "@ts-rest/core";
-import { command, computed, type Command, type Computed } from "ccstate";
+import { computed, type Computed } from "ccstate";
 import type { z } from "zod";
 
 import { badRequest, badRequestMessage } from "../../lib/error";
@@ -116,31 +116,36 @@ export function queryOf<T>(_route: RouteWithQuery<T>): Computed<T> {
   return validatedQuery$ as Computed<T>;
 }
 
+const bodyResult$ = computed(async (get): Promise<BodyResult<unknown>> => {
+  const text = await get(request$).text();
+
+  const parsed = text.length === 0 ? {} : safeJsonParse(text);
+  if (parsed === undefined) {
+    return {
+      ok: false,
+      response: badRequestMessage("Invalid JSON in request body"),
+    };
+  }
+
+  const route = get(route$);
+  const schema = "body" in route ? route.body : undefined;
+  if (!isZodLikeSchema(schema)) {
+    return { ok: true, data: parsed };
+  }
+
+  const result = schema.safeParse(parsed);
+  if (!result.success) {
+    return {
+      ok: false,
+      response: badRequest(result.error?.issues[0] ?? FALLBACK_ISSUE),
+    };
+  }
+
+  return { ok: true, data: result.data };
+});
+
 export function bodyResultOf<T>(
-  route: RouteWithBody<T>,
-): Command<Promise<BodyResult<T>>, [AbortSignal]> {
-  return command(
-    async ({ get }, signal: AbortSignal): Promise<BodyResult<T>> => {
-      const text = await get(request$).text();
-      signal.throwIfAborted();
-
-      const parsed = text.length === 0 ? {} : safeJsonParse(text);
-      if (parsed === undefined) {
-        return {
-          ok: false,
-          response: badRequestMessage("Invalid JSON in request body"),
-        };
-      }
-
-      const result = route.body.safeParse(parsed);
-      if (!result.success) {
-        return {
-          ok: false,
-          response: badRequest(result.error.issues[0] ?? FALLBACK_ISSUE),
-        };
-      }
-
-      return { ok: true, data: result.data };
-    },
-  );
+  _route: RouteWithBody<T>,
+): Computed<Promise<BodyResult<T>>> {
+  return bodyResult$ as Computed<Promise<BodyResult<T>>>;
 }
