+++ turbo/apps/api/src/signals/external/env.ts
@@ -2,8 +2,6 @@ import { computed, type Computed } from "ccstate";
 
 import { env, type EnvKey } from "../../lib/env";
 
-export { env, type EnvKey } from "../../lib/env";
-
 function envComputed<K extends EnvKey>(
   name: K,
 ): Computed<ReturnType<typeof env<K>>> {
