--- turbo/apps/api/src/signals/external/log.ts
@@ -1,1 +1,0 @@
-export { logger } from "../../lib/log";
